添加删除的方法

+++ src/main/java/com/myfirst/service/GuideService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * Created by Administrator on 2017/5/4.
@@ -22,8 +23,8 @@ public class GuideService {
     }
 
     //展示导游信息
-    public List<GuideInfo> showGuideInfo(int size,int page) {
-        return guideInfoDao.selectAllGuideInfo(page*size,size);
+    public List<GuideInfo> showGuideInfo(int size, int page) {
+        return guideInfoDao.selectAllGuideInfo(page * size, size);
     }
 
     //展示具体每一个导游信息
@@ -32,7 +33,11 @@ public class GuideService {
     }
 
     //删除导游信息
-    public int deleteGuideInfo(int travelid) {
+    public int deleteGuideInfo(int travelid, Map<String, Object> responseMap) {
+        if (null == showGuideInfoByid(travelid)) {
+            responseMap.put("error", "导游信息不存在或者已经删除,请不要重复删除！");
+            return -1;
+        }
         return guideInfoDao.deleteGuideInfoById(travelid);
     }
